near_miss_mapper/backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
motor==3.3.0
python-dotenv==1.0.0
pydantic==1.10.7
python-multipart==0.0.6
gunicorn==20.1.0
geojson==3.0.1
pytest==7.3.1
httpx==0.24.0
